Extract app intents into their own plugin
The PWA-shortcut / share-target / note-taking dispatcher isn't a
property of daily notes — it's an app-shell concern that happens
to delegate to today's daily note. Splitting it out so future
intents resolving to other targets (an 'open last panel' shortcut,
a tag quick-add flow) can grow here without daily-notes owning
them all.

- New plugin at src/plugins/app-intents/ contributes a single
  AppEffect (appIntentsBootstrapEffect) via appEffectsFacet, which
  resolves the layout-session block and hands off to
  consumeAppIntent. Replaces the explicit call previously hardcoded
  into App.tsx after projection.start().
- daily-notes plugin no longer re-exports consumeAppIntent; the
  app-intents plugin imports appendTodayDailyBlockInStack directly.
- staticAppExtensions registers the plugin last so its effect runs
  after every other plugin's data-layer setup.
- Tests moved with the code; added a small plugin-shape test that
  asserts the AppEffect is contributed.
