+++ src/plugins/daily-notes/index.ts
@@ -82,7 +82,6 @@ export {
   appendTodayDailyBlockInStack,
   resolveCurrentDailyNoteIso,
 } from './actions.ts'
-export { consumeAppIntent } from './appIntents.ts'
 export {
   openDailyNotePicker,
   openDailyNotePickerEvent,
